Add --dump-ib-file command
Also added UMR_PROCESS_HUB which allows umr_access_vram() to access process
memory.  This allows running things like the ring decoder on buffers of memory
mapped to the process.

Fixed the ring decoders to not try and determine if buffers are mapped
or not if the "follow_ib" option is not selected.

Signed-off-by: Tom St Denis <tom.stdenis@amd.com>